src/inference/requirements.txt
tensorflow==2.12.0
boto3==1.26
numpy==1.23.5
Pillow==9.0.1
python-dotenv==1.0.0